prefer resource constructor
