+++ services/user/Host/plugin/common/src/bucket.rs
@@ -1,11 +1,12 @@
 use crate::exports::host::common::{
     client::Guest as Client,
-    store::{Database, GuestBucket},
+    store::{Database, DbMode, GuestBucket, StorageDuration},
     types::Error,
 };
 use crate::make_error;
 use crate::supervisor::bridge::database as HostDb;
 use crate::HostCommon;
+use regex::Regex;
 
 pub struct Bucket {
     bucket_id: String,
@@ -13,12 +14,6 @@ pub struct Bucket {
 }
 
 impl Bucket {
-    pub fn new(db: Database, identifier: String) -> Self {
-        let service_account = HostCommon::get_sender_app().app.unwrap();
-        let bucket_id = format!("{}:{}", service_account, identifier);
-        Self { bucket_id, db }
-    }
-
     fn get_key(&self, key: &str) -> String {
         format!("{}:{}", self.bucket_id, key)
     }
@@ -39,9 +34,61 @@ impl Bucket {
         }
         Ok(())
     }
+
+    fn validate_identifier(identifier: &str) -> Result<(), Error> {
+        // Validate identifier with regex /^[0-9a-zA-Z-]+$/
+        let valid_identifier_regex = Regex::new(r"^[0-9a-zA-Z-]+$").unwrap();
+        if !valid_identifier_regex.is_match(&identifier) {
+            return Err(make_error(
+                "Invalid bucket identifier: Identifier must conform to /^[0-9a-zA-Z-]+$/",
+            )
+            .into());
+        }
+        Ok(())
+    }
+
+    // +---------------------+---------------------+---------------------+
+    // |                     |  NonTransactional   |    Transactional*   |
+    // +---------------------+---------------------+---------------------+
+    // | Ephemeral*          |      Valid          |      Not valid      |
+    // +---------------------+---------------------+---------------------+
+    // | Session*            |      Valid          |        Valid        |
+    // +---------------------+---------------------+---------------------+
+    // | Persistent          |      Valid          |        Valid        |
+    // +---------------------+---------------------+---------------------+
+    //
+    // * Not yet supported
+    //
+    fn validate_db(db: &Database) -> Result<(), Error> {
+        if db.mode == DbMode::Transactional {
+            return Err(make_error("Transactional database not yet supported"));
+        }
+        if db.duration == StorageDuration::Ephemeral {
+            return Err(make_error("Transient database not yet supported"));
+        }
+        if db.duration == StorageDuration::Session {
+            return Err(make_error("Session database not yet supported"));
+        }
+        if db.duration == StorageDuration::Ephemeral {
+            if db.mode == DbMode::Transactional {
+                return Err(make_error(
+                    "Ephemeral database not supported in transactional mode",
+                ));
+            }
+        }
+        Ok(())
+    }
 }
 
 impl GuestBucket for Bucket {
+    fn new(db: Database, identifier: String) -> Self {
+        Self::validate_db(&db).unwrap();
+        Self::validate_identifier(&identifier).unwrap();
+        let service_account = HostCommon::get_sender_app().app.unwrap();
+        let bucket_id = format!("{}:{}", service_account, identifier);
+        Self { bucket_id, db }
+    }
+
     fn get(&self, key: String) -> Result<Option<Vec<u8>>, Error> {
         self.validate_key_size(&key)?;
